src/requirements.txt
aiosqlite>=0.20.0, <1.0.0
chainlit==1.3.2
httpx>=0.27.2, <1.0.0
aiohttp
openai>=1.51.0, <2.0.0
pydantic==2.10.1
pandas>=2.2.3, <3.0.0
pillow>=11.0.0, <12.0.0
python_dotenv>=1.0.1, <2.0.0
azure-identity>=1.19.0, <2.0.0
azure-ai-projects==1.0.0b4